feat: ChessGameController 인터페이스 추가
- 콘솔기반, 웹 기반으로 나누기 위해서 인터페이스를 정의하였습니다.
